+++ odoo/local-src/sf_purchase/__manifest__.py
@@ -8,7 +8,13 @@
     'website': 'https://www.sensefly.com',
     'category': 'SenseFly',
     'depends': ['purchase'],
-    'data': ['views/purchase.xml'],
+    'data': [
+        'report/sf_purchase_order_templates.xml',
+        'report/sf_purchase_order_internal.xml',
+        'report/sf_purchase_quotation_templates.xml',
+        'report/sf_purchase_reports.xml',
+        'views/purchase.xml'
+    ],
     'demo': [],
     'images': ['static/description/icon.png'],
     'installable': True,